requirements.txt
Django>=5.0
Pillow>=10.0
qrcode>=7.4
python-barcode>=0.15
reportlab>=4.0
django-filter>=24.2
django-crispy-forms>=2.3
crispy-bootstrap5>=0.7
djangorestframework>=3.15
python-dotenv>=1.0
